+++ src/Epidemic/Model/BirthDeathSampling.hs
@@ -2,13 +2,12 @@
 
 module Epidemic.Model.BirthDeathSampling
   ( configuration
-  , observedEvents
   , randomEvent
   , BDSRates(..)
   , BDSPopulation(..)
   ) where
 
-import Data.Maybe (fromJust, isJust, isNothing)
+import Data.Maybe (fromJust)
 import qualified Data.Vector as V
 import Epidemic
 import Epidemic.Types.Events
@@ -17,7 +16,6 @@ import Epidemic.Types.Population
 import Epidemic.Types.Simulation
   ( SimulationConfiguration(..)
   , SimulationRandEvent(..)
-  , SimulationState(..)
   )
 import Epidemic.Utility
 import System.Random.MWC
@@ -99,8 +97,3 @@ randomBirthDeathSamplingEvent bdsRates@(BDSRates br dr sr) currTime pop@(BDSPopu
                , BDSPopulation unselectedPeople
                , currId)
            _ -> error "no birth-death-sampling event selected."
-
--- | Just the events that appear in the reconstructed tree.
--- TODO Implement this!!!
-observedEvents :: [EpidemicEvent] -> Maybe [EpidemicEvent]
-observedEvents = undefined