+++ src/Epidemic/Model/LogisticBDSD.hs
@@ -3,7 +3,6 @@
 
 module Epidemic.Model.LogisticBDSD
   ( configuration
-  , observedEvents
   , randomEvent
   , LogisticBDSDParameters(..)
   , LogisticBDSDPopulation(..)
@@ -35,7 +34,6 @@ import Epidemic.Types.Population
 import Epidemic.Types.Simulation
   ( SimulationConfiguration(..)
   , SimulationRandEvent(..)
-  , SimulationState(..)
   )
 import Epidemic.Utility
   ( initialIdentifier
@@ -170,10 +168,3 @@ randomDisasterEvent (dsstrTime, dsstrProb) (LogisticBDSDPopulation (People currP
    in return
         ( Disaster dsstrTime (People sampledPeople)
         , LogisticBDSDPopulation (People unsampledPeople))
-
--- | From a full list of epidemic events that occurred in a simulation return
--- just the ones that were observed. For the logistic-BDSD this is those that
--- are part of the reconstructed tree and those that occurred in disasters. If
--- this is not possible return an error message in a string.
-observedEvents :: [EpidemicEvent] -> Either String [EpidemicEvent]
-observedEvents = undefined