+++ src/Epidemic/Types/Events.hs
@@ -3,14 +3,11 @@
 
 module Epidemic.Types.Events
   ( EpidemicEvent(Infection, Removal, Sampling, Catastrophe,
-              Occurrence, Disaster)
+              Occurrence, Disaster, Extinction, StoppingTime)
   , EpidemicTree(Branch, Leaf, Shoot)
   , maybeEpidemicTree
+  , isExtinctionOrStopping
   , eventTime
-  , ReconstructedTree(RBranch, RLeaf)
-  , maybeReconstructedTree
-  , PointProcessEvents(PointProcessEvents)
-  , pointProcessEvents
   , derivedFrom
   , Newick
   , asNewickString
@@ -29,16 +26,29 @@ import GHC.Generics
 -- | Events that can occur in an epidemic with their absolute time.
 data EpidemicEvent
   = Infection AbsoluteTime Person Person -- ^ infection time, infector, infectee
-  | Removal AbsoluteTime Person          -- ^ removal without observation
-  | Sampling AbsoluteTime Person         -- ^ removal and inclusion in phylogeny
-  | Catastrophe AbsoluteTime People      -- ^ scheduled sampling of lineages
-  | Occurrence AbsoluteTime Person       -- ^ removal and observed by not in phylogeny
-  | Disaster AbsoluteTime People         -- ^ scheduled occurrence of lineages
+  | Removal AbsoluteTime Person -- ^ removal without observation
+  | Sampling AbsoluteTime Person -- ^ removal and inclusion in phylogeny
+  | Catastrophe AbsoluteTime People -- ^ scheduled sampling of lineages
+  | Occurrence AbsoluteTime Person -- ^ removal and observed by not in phylogeny
+  | Disaster AbsoluteTime People -- ^ scheduled occurrence of lineages
+  | Extinction -- ^ epidemic went extinct time time can be recovered from the preceeding removal
+  | StoppingTime -- ^ the simulation reached the stopping time
   deriving (Show, Generic, Eq)
 
 instance Json.FromJSON EpidemicEvent
 
 instance Json.ToJSON EpidemicEvent
+
+-- | A representation of the whole transmission tree in a realisation of an
+-- epidemic including the unobserved leaves. Lineages that are still extant are
+-- modelled as @Shoots@ and contain a `Person` as their data rather than an
+-- event since definition they do not have an event associated with them as a
+-- leaf.
+data EpidemicTree
+  = Branch EpidemicEvent EpidemicTree EpidemicTree -- ^ Internal node representing infection event
+  | Leaf EpidemicEvent -- ^ External node representing removal event which could be observed or not
+  | Shoot Person -- ^ External node representing extant lineages
+  deriving (Show, Eq)
 
 instance Csv.ToRecord EpidemicEvent where
   toRecord e =
@@ -60,6 +70,17 @@ instance Csv.ToRecord EpidemicEvent where
         Csv.record ["occurrence", Csv.toField time, Csv.toField person, "NA"]
       (Disaster time people) ->
         Csv.record ["disaster", Csv.toField time, Csv.toField people, "NA"]
+      Extinction -> Csv.record ["extinction", "NA", "NA", "NA"]
+      StoppingTime -> Csv.record ["stop", "NA", "NA", "NA"]
+
+-- | Predicate for whether an @EpidemicEvent@ is one of the terminal events of
+-- extinction or the stopping time having been reached.
+isExtinctionOrStopping :: EpidemicEvent -> Bool
+isExtinctionOrStopping e =
+  case e of
+    Extinction -> True
+    StoppingTime -> True
+    _ -> False
 
 et :: B.ByteString -> Csv.Record -> Bool
 et bs r = (== bs) . head $ V.toList r
@@ -69,16 +90,26 @@ instance Csv.FromRecord EpidemicEvent where
     | et "infection" r =
       Infection <$> (r Csv..! 1) <*> (Person <$> Identifier <$> (r Csv..! 2)) <*>
       (Person <$> (r Csv..! 3))
-    | et "removal" r = Removal <$> (r Csv..! 1) <*> (Person <$> Identifier <$> (r Csv..! 2))
-    | et "sampling" r = Sampling <$> (r Csv..! 1) <*> (Person <$> Identifier <$> (r Csv..! 2))
+    | et "removal" r =
+      Removal <$> (r Csv..! 1) <*> (Person <$> Identifier <$> (r Csv..! 2))
+    | et "sampling" r =
+      Sampling <$> (r Csv..! 1) <*> (Person <$> Identifier <$> (r Csv..! 2))
     | et "catastrophe" r = Catastrophe <$> (r Csv..! 1) <*> (r Csv..! 2)
     | et "occurrence" r =
       Occurrence <$> (r Csv..! 1) <*> (Person <$> Identifier <$> (r Csv..! 2))
     | et "disaster" r = Disaster <$> (r Csv..! 1) <*> (r Csv..! 2)
     | otherwise = undefined
 
--- | Epidemic Events are ordered based on which occurred first.
+-- | Epidemic Events are ordered based on which occurred first. Since
+-- 'Extinction' and 'StoppingTime' events are there as placeholders they are
+-- placed as the end of the order.
 instance Ord EpidemicEvent where
+  Extinction <= Extinction = True
+  Extinction <= StoppingTime = True
+  Extinction <= _ = False
+  StoppingTime <= Extinction = False
+  StoppingTime <= StoppingTime = True
+  StoppingTime <= _ = False
   e1 <= e2 = eventTime e1 <= eventTime e2
 
 -- | The absolute time an event occurred.
@@ -91,68 +122,80 @@ eventTime e =
     Catastrophe time _ -> time
     Occurrence time _ -> time
     Disaster time _ -> time
+    _ -> undefined
 
 -- | The events that occurred as a result of the existance of the given person.
-derivedFrom :: Person
-            -> [EpidemicEvent]  -- ^ ordered epidemic events
-            -> [EpidemicEvent]
+derivedFrom ::
+     Person
+  -> [EpidemicEvent] -- ^ ordered epidemic events
+  -> [EpidemicEvent]
 derivedFrom person = derivedFromPeople (asPeople [person])
 
 -- | The events that occurred as a result of the existance of a group of people
-derivedFromPeople :: People
-                  -> [EpidemicEvent]  -- ^ ordered epidemic events
-                  -> [EpidemicEvent]
+derivedFromPeople ::
+     People
+  -> [EpidemicEvent] -- ^ ordered epidemic events
+  -> [EpidemicEvent]
 derivedFromPeople _ [] = []
-derivedFromPeople people (e:es) = case e of
-  Infection _ p1 p2 -> if includesPerson people p1 || includesPerson people p2
-                       then let people' = addPerson p2 (addPerson p1 people)
-                             in e : derivedFromPeople people' es
-                       else derivedFromPeople people es
-  Removal _ p -> let derivedEvents = derivedFromPeople people es
-                  in if includesPerson people p
-                     then e:derivedEvents
-                     else derivedEvents
-  Sampling _ p -> let derivedEvents = derivedFromPeople people es
-                   in if includesPerson people p
-                      then e:derivedEvents
-                      else derivedEvents
-  Catastrophe _ ps -> let derivedEvents = derivedFromPeople people es
-                       in if haveCommonPeople people ps
-                          then e:derivedEvents
-                          else derivedEvents
-  Occurrence _ p -> let derivedEvents = derivedFromPeople people es
-                     in if includesPerson people p
-                        then e:derivedEvents
-                        else derivedEvents
-  Disaster _ ps -> let derivedEvents = derivedFromPeople people es
-                    in if haveCommonPeople people ps
-                       then e:derivedEvents
-                       else derivedEvents
+derivedFromPeople people (e:es) =
+  case e of
+    Infection _ p1 p2 ->
+      if includesPerson people p1 || includesPerson people p2
+        then let people' = addPerson p2 (addPerson p1 people)
+              in e : derivedFromPeople people' es
+        else derivedFromPeople people es
+    Removal _ p ->
+      let derivedEvents = derivedFromPeople people es
+       in if includesPerson people p
+            then e : derivedEvents
+            else derivedEvents
+    Sampling _ p ->
+      let derivedEvents = derivedFromPeople people es
+       in if includesPerson people p
+            then e : derivedEvents
+            else derivedEvents
+    Catastrophe _ ps ->
+      let derivedEvents = derivedFromPeople people es
+       in if haveCommonPeople people ps
+            then e : derivedEvents
+            else derivedEvents
+    Occurrence _ p ->
+      let derivedEvents = derivedFromPeople people es
+       in if includesPerson people p
+            then e : derivedEvents
+            else derivedEvents
+    Disaster _ ps ->
+      let derivedEvents = derivedFromPeople people es
+       in if haveCommonPeople people ps
+            then e : derivedEvents
+            else derivedEvents
+    _ -> [e]
 
-{-| A representation of the whole transmission tree in a realisation of an
-epidemic including the unobserved leaves. Lineages that are still extant are
-modelled as shoots and contain a `Person` as their data rather than an event.
--}
-data EpidemicTree
-  = Branch EpidemicEvent EpidemicTree EpidemicTree -- ^ Internal node representing infection event
-  | Leaf EpidemicEvent                             -- ^ External node representing removal event
-  | Shoot Person                                   -- ^ External node representing extant lineages
-  deriving (Show,Eq)
-
-
--- | A tree representation of the epidemic events.
-maybeEpidemicTree :: [EpidemicEvent] -- ^ ordered epidemic events
-                  -> Maybe EpidemicTree
-maybeEpidemicTree [] = Nothing
-maybeEpidemicTree [e] = case e of
-  Catastrophe _ people -> if nullPeople people
-                          then Nothing
-                          else Just (Leaf e)
-  Disaster _ people -> if nullPeople people
-                       then Nothing
-                       else Just (Leaf e)
-  Infection _ p1 p2 -> Just (Branch e (Shoot p1) (Shoot p2))
-  _ -> Just (Leaf e)
+-- | A tree representation of /all/ the epidemic events including those that
+-- were not observed. The scheduled observations can return a 'Left' or just
+-- proceed to the next epidemic event if there were no people sampled. This is
+-- because in these cases the event had no effect upon the transmission tree. If
+-- there were multiple people sampled in a scheduled event, then this event will
+-- appear in multiple leaves since each leaf is due to this event.
+maybeEpidemicTree ::
+     [EpidemicEvent] -- ^ ordered epidemic events
+  -> Either String EpidemicTree
+maybeEpidemicTree [] =
+  Left "There are no EpidemicEvents to construct a tree with."
+maybeEpidemicTree [e] =
+  case e of
+    Catastrophe _ people ->
+      if nullPeople people
+        then Left "The last event is a catastrophe with no people sampled"
+        else Right (Leaf e)
+    Disaster _ people ->
+      if nullPeople people
+        then Left "The last event is a disaster with no people sampled"
+        else Right (Leaf e)
+    Infection _ p1 p2 -> Right (Branch e (Shoot p1) (Shoot p2))
+    Extinction -> Left "Extinction event encountered"
+    StoppingTime -> Left "Stopping time encountered"
+    _ -> Right (Leaf e)
 maybeEpidemicTree (e:es:ess) =
   case e of
     Infection _ p1 p2 ->
@@ -160,98 +203,29 @@ maybeEpidemicTree (e:es:ess) =
           infecteeEvents = derivedFrom p2 (es : ess)
        in do leftTree <-
                if null infectorEvents
-               then Just (Shoot p1)
-               else maybeEpidemicTree infectorEvents
+                 then Right (Shoot p1)
+                 else maybeEpidemicTree infectorEvents
              rightTree <-
                if null infecteeEvents
-               then Just (Shoot p2)
-               else maybeEpidemicTree infecteeEvents
+                 then Right (Shoot p2)
+                 else maybeEpidemicTree infecteeEvents
              return $ Branch e leftTree rightTree
-    Catastrophe _ people -> if nullPeople people
-                            then maybeEpidemicTree (es:ess)
-                            else Just (Leaf e)
-    Disaster _ people -> if nullPeople people
-                         then maybeEpidemicTree (es:ess)
-                         else Just (Leaf e)
-    _ -> Just (Leaf e)
+    Catastrophe _ people ->
+      if nullPeople people
+        then maybeEpidemicTree (es : ess)
+        else Right (Leaf e)
+    Disaster _ people ->
+      if nullPeople people
+        then maybeEpidemicTree (es : ess)
+        else Right (Leaf e)
+    _ -> Right (Leaf e)
 
-{-| A representation of the reconstructed tree which is the phylogeny connecting
-  all the `Sampling` and `Catastrophe` events.
--}
-data ReconstructedTree
-  = RBranch EpidemicEvent ReconstructedTree ReconstructedTree
-  | RLeaf EpidemicEvent
-  deriving (Show, Eq)
-
-
--- | A tree representation of the reconstructed phylogeny.
-maybeReconstructedTree :: EpidemicTree -> Maybe ReconstructedTree
-
-maybeReconstructedTree Shoot{} = Nothing
-
-maybeReconstructedTree (Leaf e) = case e of
-  Sampling{} -> Just $ RLeaf e
-  Catastrophe{} -> Just $ RLeaf e
-  _ -> Nothing
-
-maybeReconstructedTree (Branch e@Infection{} lt rt)
-  | hasSequencedLeaf lt && hasSequencedLeaf rt =
-    do
-      rlt <- maybeReconstructedTree lt
-      rrt <- maybeReconstructedTree rt
-      Just $ RBranch e rlt rrt
-  | hasSequencedLeaf lt = maybeReconstructedTree lt
-  | hasSequencedLeaf rt = maybeReconstructedTree rt
-  | otherwise = Nothing
-maybeReconstructedTree Branch{} = Nothing
-
--- | Predicate for whether an `EpidemicTree` has a leaf which corresponds to a
--- node in the `ReconstructedTree`.
-hasSequencedLeaf :: EpidemicTree -> Bool
-
-hasSequencedLeaf Shoot {} = False
-
-hasSequencedLeaf (Leaf e) =
-  case e of
-    Sampling {} -> True
-    Catastrophe {} -> True
-    _ -> False
-
-hasSequencedLeaf (Branch _ lt rt) = hasSequencedLeaf lt || hasSequencedLeaf rt
-
-
-{-| A representation of the events that can be observed in an epidemic but which
-  are not included in the reconstructed tree, i.e. the `Occurrence` and
-  `Disaster` events.
--}
-newtype PointProcessEvents = PointProcessEvents [EpidemicEvent]
-
--- | Extract the events from an epidemic tree which are observed but not part of
--- the reconstructed tree.
-pointProcessEvents :: EpidemicTree -> PointProcessEvents
-
-pointProcessEvents Shoot {} = PointProcessEvents []
-
-pointProcessEvents (Leaf e) = case e of
-  Occurrence {} -> PointProcessEvents [e]
-  Disaster {} -> PointProcessEvents [e]
-  _ -> PointProcessEvents []
-
-pointProcessEvents (Branch _ lt rt) =
-  let (PointProcessEvents lEs) = pointProcessEvents lt
-      (PointProcessEvents rEs) = pointProcessEvents rt
-      allEs = List.sort $ lEs ++ rEs
-      in PointProcessEvents allEs
-
-class Newick t where
-  -- | Return a representation of the tree in Newick format.
-  asNewickString :: (AbsoluteTime,Person) -- ^ The person and time of the start of the tree
-                 -> t
-                 -> Maybe (BBuilder.Builder, [EpidemicEvent])
-
-
-ampersandBuilder :: BBuilder.Builder
-ampersandBuilder = BBuilder.charUtf8 '&'
+class Newick t
+  where
+  asNewickString ::
+       (AbsoluteTime, Person) -- ^ The person and time of the start of the tree
+    -> t
+    -> Maybe (BBuilder.Builder, [EpidemicEvent])
 
 colonBuilder :: BBuilder.Builder
 colonBuilder = BBuilder.charUtf8 ':'
@@ -265,12 +239,6 @@ rightBraceBuilder = BBuilder.charUtf8 ')'
 commaBuilder :: BBuilder.Builder
 commaBuilder = BBuilder.charUtf8 ','
 
-catastrophePeopleBuilder :: People -> BBuilder.Builder
-catastrophePeopleBuilder (People persons) =
-  mconcat $
-  List.intersperse ampersandBuilder [personByteString p | p <- V.toList persons]
-
-
 instance Newick EpidemicTree where
   asNewickString (_, p) (Shoot p') =
     if p /= p'
@@ -278,33 +246,35 @@ instance Newick EpidemicTree where
       else let identifier = personByteString p
                branchLength = BBuilder.stringUtf8 "Infinity"
             in Just (identifier <> colonBuilder <> branchLength, [])
-
   asNewickString (t, p) (Leaf e) =
     let identifier = personByteString p
-        branchLength a b = BBuilder.doubleDec td where (TimeDelta td) = timeDelta a b
-    in case e of
-      Infection {} -> Nothing
-      (Removal t' p') ->
-        if p /= p'
-        then Nothing
-        else Just (identifier <> colonBuilder <> branchLength t t', [e])
-      (Sampling t' p') ->
-        if p /= p'
-        then Nothing
-        else Just (identifier <> colonBuilder <> branchLength t t', [e])
-      (Catastrophe t' ps) ->
-        if ps `includesPerson` p
-        then Just (identifier <> colonBuilder <> branchLength t t', [e])
-        else Nothing
-      (Occurrence t' p') ->
-        if p /= p'
-        then Nothing
-        else Just (identifier <> colonBuilder <> branchLength t t', [e])
-      (Disaster t' ps) ->
-        if ps `includesPerson` p
-        then Just (identifier <> colonBuilder <> branchLength t t', [e])
-        else Nothing
-
+        branchLength a b = BBuilder.doubleDec td
+          where
+            (TimeDelta td) = timeDelta a b
+     in case e of
+          Infection {} -> Nothing
+          (Removal t' p') ->
+            if p /= p'
+              then Nothing
+              else Just (identifier <> colonBuilder <> branchLength t t', [e])
+          (Sampling t' p') ->
+            if p /= p'
+              then Nothing
+              else Just (identifier <> colonBuilder <> branchLength t t', [e])
+          (Catastrophe t' ps) ->
+            if ps `includesPerson` p
+              then Just (identifier <> colonBuilder <> branchLength t t', [e])
+              else Nothing
+          (Occurrence t' p') ->
+            if p /= p'
+              then Nothing
+              else Just (identifier <> colonBuilder <> branchLength t t', [e])
+          (Disaster t' ps) ->
+            if ps `includesPerson` p
+              then Just (identifier <> colonBuilder <> branchLength t t', [e])
+              else Nothing
+          Extinction -> Nothing
+          StoppingTime -> Nothing
   asNewickString (t, p) (Branch e lt rt) =
     case e of
       (Infection t' p1 p2) ->
@@ -313,37 +283,13 @@ instance Newick EpidemicTree where
           else do
             (leftNS, leftEs) <- asNewickString (t', p1) lt
             (rightNS, rightEs) <- asNewickString (t', p2) rt
-            let branchLength = BBuilder.doubleDec td where (TimeDelta td) = timeDelta t t'
+            let branchLength = BBuilder.doubleDec td
+                  where
+                    (TimeDelta td) = timeDelta t t'
             return
               ( leftBraceBuilder <>
                 leftNS <>
-                commaBuilder <> rightNS <> rightBraceBuilder <> colonBuilder <> branchLength
+                commaBuilder <>
+                rightNS <> rightBraceBuilder <> colonBuilder <> branchLength
               , List.sort $ leftEs ++ rightEs)
-      _ -> Nothing
-
-
-
-instance Newick ReconstructedTree where
-  asNewickString (t, _) (RLeaf e) =
-    let branchLength a b = BBuilder.doubleDec td where (TimeDelta td) = timeDelta a b
-    in case e of
-      (Sampling t' p) -> Just ((personByteString p) <> colonBuilder <> branchLength t t', [e])
-      Infection {} -> Nothing
-      Removal {} -> Nothing
-      (Catastrophe t' ps) -> Just (catastrophePeopleBuilder ps <> colonBuilder <> branchLength t t', [e])
-      Occurrence {} -> Nothing
-      Disaster {} -> Nothing
-
-  asNewickString (t, _) (RBranch e lt rt) =
-    case e of
-      (Infection t' p1 p2) ->
-        do
-          (leftNS, leftEs) <- asNewickString (t', p1) lt
-          (rightNS, rightEs) <- asNewickString (t', p2) rt
-          let branchLength = BBuilder.doubleDec td where (TimeDelta td) = timeDelta t t'
-          return
-            ( leftBraceBuilder <>
-              leftNS <>
-              commaBuilder <> rightNS <> rightBraceBuilder <> colonBuilder <> branchLength
-            , List.sort $ leftEs ++ rightEs)
       _ -> Nothing